Modufications of algorithms

+++ rawTiffStructureToDen.py
@@ -25,12 +25,13 @@ parser = argparse.ArgumentParser()
 parser.add_argument("inputh5")
 parser.add_argument("outputDir")
 parser.add_argument(
-    "--raw-dir",
-    default=None,
-    type=str,
-    help=
-    "Provide raw directory where to find files, by default parrent directory of inputh5."
+	"--raw-dir",
+	default=None,
+	type=str,
+	help=
+	"Provide raw directory where to find files, by default parrent directory of inputh5."
 )
+parser.add_argument("--fix-corrupted-h5", action="store_true", help="Fix corrupted HDF5 file by scanning for TIFF files.")
 parser.add_argument("--force", action="store_true")
 
 try:
@@ -61,14 +62,16 @@ def insertToDf(df, dat, name):
 
 #To write dataframe to den
 def writeDenFile(df, inputDir, denFile, force=False):
+	if df.empty:
+		raise ValueError("Dataframe is empty, can not write file %s" % (os.path.realpath(denFile)))
 	if os.path.exists(denFile):
 		if force:
 			os.remove(denFile)
 		else:
 			raise IOError("File %s exists, add force to overwrite" % (denFile))
-	df = df.sort_values("time", ascending=True)
-	fileStr = df["image_file"].iloc[
-	    0]  #In some versions this is string but in some bytes
+	if "time" in df.columns:
+		df = df.sort_values("time", ascending=True)
+	fileStr = df["image_file"].iloc[0]#In some versions this is string but in some bytes
 	if isinstance(fileStr, bytes):
 		fileStr = fileStr.decode("utf-8")
 	file = os.path.join(inputDir, fileStr.lstrip("/"))
@@ -78,23 +81,34 @@ def writeDenFile(df, inputDir, denFile, force=False):
 	dimx = img.shape[1]
 	dimz = len(df)
 	dtype = img.dtype
-	#    img=TIFF.open(file)
-	#    img = img.read_image()
-	#    dimy=img.shape[0]
-	#    dimx=img.shape[1]
+	#	 img=TIFF.open(file)
+	#	 img = img.read_image()
+	#	 dimy=img.shape[0]
+	#	 dimx=img.shape[1]
 	print("Creating %s %dx%dx%d" % (denFile, dimx, dimy, dimz))
 	DEN.writeEmptyDEN(denFile, [dimx, dimy, dimz], force=True)
 	for i in range(len(df)):
-		fileStr = df["image_file"].iloc[
-		    i]  #In some versions this is string but in some bytes
+		fileStr = df["image_file"].iloc[i]#In some versions this is string but in some bytes
 		if isinstance(fileStr, bytes):
 			fileStr = fileStr.decode("utf-8")
 		img = np.array(Image.open(os.path.join(inputDir, fileStr.lstrip("/"))),
-		               dtype=np.float32)
+					   dtype=np.float32)
 		print("Writing %d-th file %s of shape %d,%d into %s" %
-		      (i, fileStr, img.shape[0], img.shape[1], denFile))
+			  (i, fileStr, img.shape[0], img.shape[1], denFile))
 		DEN.writeFrame(denFile, i, img, force=True)
 
+# Function to scan directory for TIFF files and create a DataFrame
+def scanForTiffFiles(directory, exclude_files):
+	tiff_files = []
+	for root, _, files in os.walk(directory):
+		for file in files:
+			if file.lower().endswith('.tiff') or file.lower().endswith('.tif'):
+				filepath = os.path.join(root, file)
+				relative_path = os.path.relpath(filepath, directory)
+				if relative_path not in exclude_files:
+					tiff_files.append(relative_path)
+	tiff_files.sort()
+	return pd.DataFrame({'image_file': tiff_files})
 
 #Parse input data
 if ARG.raw_dir is not None:
@@ -106,26 +120,6 @@ if not os.path.exists(ARG.outputDir):
 	print("Create dir %s" % (ARG.outputDir))
 	os.mkdir(ARG.outputDir)
 
-#f = h5py.File(ARG.inputH5, 'r')
-#data = f["entry/scan/data"]
-#Now each subfolder has two datasets (time, value)
-#Time is uint64 and has a representation of 'datetime64[ms]'
-#Extract all the information into dtuple list
-#labels = list(data.keys())
-#if len(labels) < 1:
-#    sys.exit("Error: labels count is %d!"%(labels.count))
-#print(data["image_file/time"].value)
-#dataset.value attribute was removed after h5py 2.9.0
-#df = pd.DataFrame(columns=labels+["time"], index=list(data["%s/time"%(labels[0])][()]))
-
-#for ind in df.index:
-#    df.loc[ind]["time"] = pd.to_datetime(ind, unit="ms")
-#for lab in labels:
-#    insertToDf(df, data, lab)
-#imageFiles=f["entry/scan/data/image_file/value"].value
-#imageTimes=np.array(f["entry/scan/data/image_file/time"].value, dtype='datetime64[ms]')
-#Most likely this is np.datetime64[ms]
-#Dark fields, white fields, scan
 df = PETRA.scanDataset(ARG.inputh5)
 
 dark = df.loc[df["image_key"] == 2]
@@ -134,16 +128,12 @@ scan = df.loc[df["image_key"] == 0]
 
 writeDenFile(dark, inputDir, os.path.join(ARG.outputDir, "dar.den"), ARG.force)
 writeDenFile(white, inputDir, os.path.join(ARG.outputDir, "ref.den"), ARG.force)
-writeDenFile(scan, inputDir, os.path.join(ARG.outputDir, "img.den"), ARG.force)
-#Now create reduced scan frame
-#tst = scan.head(2)
-#writeDenFile(tst, inputDir, os.path.join(ARG.outputDir, "x.den"), ARG.force)
 
-#
-#
-#
-#
-#
-#scan.sort_values("time", ascending=True)
-#for f in scan["image_file"]:
-#    print(f)
+if ARG.fix_corrupted_h5 and scan.empty:
+	print("HDF5 file is corrupted. Scanning directory for TIFF files to create img.den...")
+	dark_files = set(dark["image_file"])
+	white_files = set(white["image_file"])
+	exclude_files = dark_files.union(white_files)
+	scan = scanForTiffFiles(inputDir, exclude_files)
+
+writeDenFile(scan, inputDir, os.path.join(ARG.outputDir, "img.den"), ARG.force)
